+++ src/Renderers/Renderer.php
@@ -1,0 +1,11 @@
+<?php
+/**
+* Abstract class for renderers definition
+*/
+namespace PHPForm\Renderers;
+
+abstract class Renderer
+{
+    abstract public function getTemplate($template_name);
+    abstract public function render($template_name, $context);
+}