+++ src/Widgets/Input.php
@@ -4,24 +4,10 @@
  */
 namespace PHPForm\Widgets;
 
-use PHPForm\PHPFormConfig;
-
 abstract class Input extends Widget
 {
-    /**
-    * @var string The input type to use for the widget.
-    */
-    protected $input_type = null;
-
-    /**
-     * The constructor.
-     */
-    public function __construct(array $attrs = null)
-    {
-        $this->template = PHPFormConfig::getITemplate("INPUT");
-
-        parent::__construct($attrs);
-    }
+    const TEMPLATE = 'input.html';
+    const INPUT_TYPE = null;
 
     /**
      * Prepare context to be used on render method.
@@ -35,7 +21,9 @@ abstract class Input extends Widget
     public function getContext(string $name, $value, array $attrs = null)
     {
         $context = parent::getContext($name, $value, $attrs);
-        $context["type"] = $this->input_type;
+
+        $context["type"] = static::INPUT_TYPE;
+
         return $context;
     }
 }