+++ tests/unit/Errors/ErrorListTest.php
@@ -13,17 +13,4 @@ class ErrorListTest extends TestCase
         $expected = '<ul class="errorlist"><li>1</li><li>2</li><li>3</li></ul>';
         $this->assertEquals((string) $error, $expected);
     }
-
-    public function testAsUL()
-    {
-        $error = new ErrorList(array(1, 2, 3));
-        $expected = '<ul class="errorlist"><li>1</li><li>2</li><li>3</li></ul>';
-        $this->assertEquals($error->asUL(), $expected);
-    }
-
-    public function testAsULEmptyArray()
-    {
-        $error = new ErrorList();
-        $this->assertEquals($error->asUL(), '');
-    }
 }