+++ tests/unit/Utils/Attributes.php
@@ -1,5 +1,5 @@
 <?php
-namespace PHPForm\Unit\Validators;
+namespace PHPForm\Unit\Utils;
 
 use PHPUnit\Framework\TestCase;
 
@@ -7,14 +7,6 @@ use PHPForm\Utils\Attributes;
 
 class AttributesTest extends TestCase
 {
-    public function testFlatatt()
-    {
-        $attrs = array("name" => "name", "type" => "text");
-        $this->assertEquals(Attributes::flatatt($attrs), 'name="name" type="text"');
-        $this->assertEquals(Attributes::flatatt(array()), '');
-        $this->assertEquals(Attributes::flatatt("string"), '');
-    }
-
     public function testPrettyName()
     {
         $this->assertEquals(Attributes::prettyName("label_name"), 'Label name');